version history is shifted to individual view
